% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{pali_alphabet}
\alias{pali_alphabet}
\title{Pali alphabet in order}
\format{
The Pali alphabet in traditional order.
}
\usage{
pali_alphabet
}
\description{
Pali alphabet in order
}
\examples{
# Returns TRUE because a comes before b in Pali:
match("a", pali_alphabet) < match("b", pali_alphabet)
# Returns FALSE beceause c comes before b in Pali
match("b", pali_alphabet) < match("c", pali_alphabet)

}
\keyword{datasets}
