% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pali-sort.R
\name{pali_eq}
\alias{pali_eq}
\title{Equal (==) comparison function for Pali words}
\usage{
pali_eq(word1, word2)
}
\arguments{
\item{word1}{A first Pali word as a string}

\item{word2}{A second Pali word as a string}
}
\value{
TRUE if word1 and word2 are the same
}
\description{
Note that all Pali string comparisons are case-insensitive.
}
