% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pali-sort.R
\name{pali_gt}
\alias{pali_gt}
\title{Greater-than (>) comparison function for Pali words}
\usage{
pali_gt(word1, word2)
}
\arguments{
\item{word1}{A first Pali word as a string}

\item{word2}{A second Pali word as a string}
}
\value{
TRUE if word1 comes after word2 alphabetically
}
\description{
Note that all Pali string comparisons are case-insensitive.
#' Also non-Pali characters are placed at the end of the
alphabet and are considered equivalent to each other.
}
