% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pali-sort.R
\name{pali_string_fix}
\alias{pali_string_fix}
\title{Make built-in Pali strings more readable.}
\usage{
pali_string_fix(string)
}
\arguments{
\item{string}{A Pali string that has Unicode escape characters (optional)}
}
\value{
The string converted to Pali characters (if provided)
}
\description{
CRAN requires that all Pali strings be rendered in ASCII
characters, so letters with diacritics have to be represented
with Unicode escape codes (ie, "ā" becomes "\\u0101"). This
converts all those strings back to more readable forms. If
you call it with a string argument, it returns that string
with the Unicode escape codes converted to Pali characters.
If you call without any arguments, it simply converts the
built-in strings to more readble form.
}
