% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{sati_sutta_long}
\alias{sati_sutta_long}
\title{Mahﾄ《atipatthﾄ］a Sutta in "long" form}
\format{
An object of class \code{data.frame} with 832 rows and 4 columns.
}
\source{
Vipassana Research Institute, CST4, April 2020
}
\usage{
sati_sutta_long
}
\description{
The Mahﾄ《atipatthﾄ］a Sutta or Discourse on the Establishing
of Mindfulness in "long" form.
}
\keyword{datasets}
