% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{sati_sutta_raw}
\alias{sati_sutta_raw}
\title{Mahﾄ《atipatthﾄ］a Sutta text in raw form}
\format{
A tibble with the variable:
\describe{
\item{text}{Complete text}
}
}
\source{
Vipassana Research Institute, CST4, April 2020
}
\usage{
sati_sutta_raw
}
\description{
The unprocessed text of the Mahﾄ《atipatthﾄ］a Sutta
}
\keyword{datasets}
