% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{tipitaka_names}
\alias{tipitaka_names}
\title{Names of each book of the Tipitaka, both abbreviated and
in full. These are easier to read if you call \code{pali_string_fix() first}.}
\format{
A tibble with the variables:
\describe{
  \item{book}{Abbreviated title}
  \item{name}{Full title}
}
}
\usage{
tipitaka_names
}
\description{
Names of each book of the Tipitaka, both abbreviated and
in full. These are easier to read if you call \code{pali_string_fix() first}.
}
\examples{
# Clean up the Unicode characters to make things more readble:
tipitaka_names$name <-
  stringi::stri_unescape_unicode(tipitaka_names$name)

}
\keyword{datasets}
