% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{tipitaka_raw}
\alias{tipitaka_raw}
\title{Tipitaka text in raw form}
\format{
A tibble with the variables:
\describe{
\item{text}{Text of each Tipitaka volume}
\item{book}{Abbreviated book name of each volume}
}
}
\source{
Vipassana Research Institute, CST4, April 2020
}
\usage{
tipitaka_raw
}
\description{
The unprocessed text of the Tipitaka, with one row per volume.
}
\keyword{datasets}
