% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{abhidhamma_pitaka}
\alias{abhidhamma_pitaka}
\title{All the books of the Abhidhamma Pitaka}
\format{
A tibble with the variables:
\describe{

  \item{book}{Abbreviated title}
  \item{name}{Full title}
}\
}
\usage{
abhidhamma_pitaka
}
\description{
A subset of tipitaka_names consisting of only the books of
the Abhidhamma Pitaka. These are easier to read if you call
\code{pali_string_fix() first}.
}
\examples{
# Clean up the Unicode characters to make things more readble:
abhidhamma_pitaka$name <-
  stringi::stri_unescape_unicode(abhidhamma_pitaka$name)

# Count all the words in the Abhidhamma Pitaka:
sum(tipitaka_long[tipitaka_long$book \%in\% abhidhamma_pitaka$book, "n"])

}
\keyword{datasets}
