% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{abhidhamma_pitaka}
\alias{abhidhamma_pitaka}
\title{All the books of the Abhidhamma Pitaka}
\format{
A tibble with the variables:
\describe{

  \item{book}{Abbreviated title}
  \item{name}{Full title}
}
}
\usage{
abhidhamma_pitaka
}
\description{
A subset of tipitaka_names consisting of only the books of
the Abhidhamma Pitaka.
}
\keyword{datasets}
