% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{pali_alphabet}
\alias{pali_alphabet}
\title{Pali alphabet in order}
\format{
The Pali alphabet in traditional order.
}
\usage{
pali_alphabet
}
\description{
Pali alphabet in order
}
\keyword{datasets}
