% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{pali_stop_words}
\alias{pali_stop_words}
\title{Tentative set of "stop words" for Pali}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 245 rows and 1 columns.
}
\source{
\url{https://dsal.uchicago.edu/dictionaries/pali/}
}
\usage{
pali_stop_words
}
\description{
A list of all declinables and particles from the PTS
Pali-English Dictionary.
}
\examples{
\donttest{
# Show top content words in the Tipitaka (excluding stop words)
content_words <- tipitaka_long[!tipitaka_long$word \%in\% pali_stop_words$word, ]
head(content_words[order(-content_words$n), ], 10)
}

}
\keyword{datasets}
