% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{pali_stop_words}
\alias{pali_stop_words}
\title{Tentative set of "stop words" for Pali}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 145 rows and 1 columns.
}
\source{
\url{https://dsalsrv04.uchicago.edu/dictionaries/pali/}
}
\usage{
pali_stop_words
}
\description{
A list of all declinables and particles from the PTS
Pali-English Dictionary.
}
\keyword{datasets}
