% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{package}
\name{tipitaka}
\alias{tipitaka-package}
\alias{tipitaka}
\title{tipitaka: A package for exploring the Pali Canon in R.}
\description{
The package tipitaka provides access to the complete Pali
Canon, or Tipitaka, from R. The Tipitaka is the canonical
scripture for Theravadin Buddhists worldwide. This package
includes the VRI (Vipassana Research Institute) Chattha
Sangayana edition along with tools for working with Pali text.
}
\details{
For a lemmatized critical edition of the Tipitaka with
sutta-level granularity, see the companion package
\pkg{tipitaka.critical}.
}
\section{Datasets}{

\itemize{
  \item tipitaka_raw: the complete text of the Tipitaka (VRI)
  \item tipitaka_long: the complete Tipitaka in "long" form (VRI)
  \item tipitaka_wide: the complete Tipitaka in "wide" form (VRI)
  \item tipitaka_names: the names of each book of the Tipitaka
  \item sutta_pitaka: the names of each volume of the Sutta Pitaka
  \item vinaya_pitaka: the names of each volume of the Vinaya Pitaka
  \item abhidhamma_pitaka: the names of each volume of the Abhidhamma Pitaka
  \item pali_alphabet: the complete Pali alphabet in traditional order
  \item pali_stop_words: a set of "stop words" for Pali
}
}

\section{Tools}{

Functions for working with Pali text:
\itemize{
  \item pali_lt: less-than function for Pali strings
  \item pali_gt: greater-than function for Pali strings
  \item pali_eq: equals function for Pali strings
  \item pali_sort: sorting function for vectors of Pali strings
}
}

\author{
\strong{Maintainer}: Dan Zigmond \email{djz@shmonk.com}

}
