% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\docType{data}
\name{tipitaka_long}
\alias{tipitaka_long}
\title{Tipitaka in "long" form}
\format{
A data frame with the variables:
\describe{
\item{word}{Pali word}
\item{n}{Number of times this word appears in this book}
\item{total}{Total number of words in this book}
\item{freq}{Frequency with which this word appears in this book}
\item{book}{Abbreviated book name}
}
}
\source{
Vipassana Research Institute, CST4, April 2020
}
\usage{
tipitaka_long
}
\description{
Every word of every volume of the Tipitaka, with one word per
volume per line. Computed from \code{tipitaka_raw} on first access.
}
\keyword{datasets}
