% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{tipitaka_names}
\alias{tipitaka_names}
\title{Names of each book of the Tipitaka, both abbreviated and
in full.}
\format{
A tibble with the variables:
\describe{
  \item{book}{Abbreviated title}
  \item{name}{Full title}
}
}
\usage{
tipitaka_names
}
\description{
Names of each book of the Tipitaka, both abbreviated and
in full.
}
\keyword{datasets}
