% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipitaka-docs.R
\docType{data}
\name{vinaya_pitaka}
\alias{vinaya_pitaka}
\title{All the books of the Vinaya Pitaka}
\format{
A tibble with the variables:
\describe{
  \item{book}{Abbreviated title}
  \item{name}{Full title}
}
}
\usage{
vinaya_pitaka
}
\description{
A subset of tipitaka_names consisting of only the books of
the Vinaya Pitaka.
}
\keyword{datasets}
